Revert to using promises and fix tests
I had a `.eslintrc.json`-file in a parent directory, which caused me to lint with a different config. Removing it revealed that I couldn't use async/await. That was what caused the tests to fail.

Reverting back to promises revealed that an exception was not caught in lib/imgAutosize.js, which is fixed now.